+++ docs/changelog.rst
@@ -4,7 +4,9 @@ Changelog
 0.2a1 - 2014-09-24
 ------------------
 
+* Fixed problem with deprecation warning (django simplejson)
 * Using python's json or simplejson
+
 
 0.1c1 - 2012-08-25
 ------------------